DEV: ensures modal has finished loading (#42)
It's unclear if it could be the cause of a timeout, but it seems like a better change to make and might give us a better error.

`<DTemplates::FilterableList>` has `loading=true` by default, so we are sure that on insert the spinner is showing. This commit ensures the spinner is gone before we consider the modal truly opened.