FIX: Slowness when listing templates and the templates category (#67)
Some communities were noticing that fetching the list of templates was very slow.
This issue wouldn't just affect the templates form, but also browsing the templates category and

Most of the communities didn't look like they were affected but in some with extreme cases, this was causing 502 errors.
The problem also seemed intermittent, with the slowness going away completely without any special action.

Upon inspecting a particularly affected instance, the problem was narrowed down to the list of subcategories from the categories templates that was repeating the same `id` multiple times.

The issue was tracked to a line that was mutating an array returned by the method `Category.subcategory_ids` and prepending an item. Unfortunately, the result of this method was cached in the `DistributedCache` and mutating the array was causing the cached value to pile up the same category `id` in the array.

This caused the database queries that used the value returned for this function to become slow over time when the argument was the templates category. The busier the Community, the worst the problem would get.

This commit removes the `prepend` that was causing the issue.